Separating UNIX and MacOSX P/Invoke calls - SCARD_READERSTATE -> using int instead of IntPtr
