+++ src/PCSC/Interop/Linux/ExtensionMethods/IntPtrExt.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace PCSC.Interop.Linux.ExtensionMethods
+{
+    internal static class IntPtrExt
+    {
+        public static IntPtr Mask(this IntPtr value, long mask) {
+            unchecked {
+                var newValue = value.ToInt64() & mask;
+
+                return new IntPtr(newValue);
+            }
+        }
+    }
+}
