+++ src/PCSC/Interop/Linux/SCARD_IO_REQUEST.cs
@@ -1,0 +1,16 @@
+using System;
+using System.Runtime.InteropServices;
+
+namespace PCSC.Interop.Linux
+{
+    [StructLayout(LayoutKind.Sequential)]
+    internal class SCARD_IO_REQUEST
+    {
+        internal SCARD_IO_REQUEST() {
+            dwProtocol = IntPtr.Zero;
+        }
+
+        internal IntPtr dwProtocol; // Protocol identifier
+        internal IntPtr cbPciLength; // Protocol Control Inf Length
+    }
+}
