+++ src/PCSC/Interop/Platform.cs
@@ -1,19 +1,30 @@
 ﻿using System;
-using PCSC.Interop.Unix;
+using PCSC.Interop.Linux;
+using PCSC.Interop.MacOSX;
 using PCSC.Interop.Windows;
 
 namespace PCSC.Interop
 {
+    internal enum PlatformType
+    {
+        Windows,
+        Linux,
+        MacOSX
+    }
+    
     /// <summary>
     /// Platform selector (Windows or UNIX)
     /// </summary>
     public static class Platform
     {
+        internal static PlatformType Type { get; }
+
+        
         /// <summary>
         /// Returns <c>true</c> if the operation system runs on Windows. <c>false</c> otherwise.
         /// </summary>
-        public static bool IsWindows { get; }
-
+        public static bool IsWindows => Type == PlatformType.Windows;
+        
         /// <summary>
         /// Platform smart card library.
         /// </summary>
@@ -22,19 +33,25 @@ namespace PCSC.Interop
         static Platform() {
             var platform = Environment.OSVersion.Platform;
 
-            if (
-                platform == PlatformID.Win32S ||
-                platform == PlatformID.Win32Windows ||
-                platform == PlatformID.Win32NT ||
-                platform == PlatformID.WinCE
-            ) {
-                IsWindows = true;
-                Lib = new WinSCardAPI();
-                return;
+            switch (platform) {
+                case PlatformID.Unix:
+                    Type = PlatformType.Linux;
+                    Lib = new PCSCliteLinux();
+                    break;
+                case PlatformID.MacOSX:
+                    Type = PlatformType.MacOSX;
+                    Lib = new PCSCliteMacOsX();
+                    break;
+                case PlatformID.Win32NT:
+                case PlatformID.Win32S:
+                case PlatformID.Win32Windows:
+                case PlatformID.WinCE:
+                    Type = PlatformType.Windows;
+                    Lib = new WinSCardAPI();
+                    break;
+                default:
+                    throw new NotSupportedException("Sorry, your OS platform is not supported.");
             }
-
-            IsWindows = false;
-            Lib = new PCSCliteAPI();
         }
     }
 }
