--- src/PCSC/Interop/Unix/SCARD_IO_REQUEST.cs
@@ -1,16 +1,0 @@
-using System;
-using System.Runtime.InteropServices;
-
-namespace PCSC.Interop.Unix
-{
-    [StructLayout(LayoutKind.Sequential)]
-    internal class SCARD_IO_REQUEST
-    {
-        internal SCARD_IO_REQUEST() {
-            dwProtocol = IntPtr.Zero;
-        }
-
-        internal IntPtr dwProtocol; // Protocol identifier
-        internal IntPtr cbPciLength; // Protocol Control Inf Length
-    }
-}
